+++ src/Commands/CreateMappedResourcesCommand.php
@@ -243,7 +243,7 @@ class CreateMappedResourcesCommand extends Command
         $input->formRequest = $this->option('with-form-request');
         $input->controllerDirectory = $this->option('controller-directory');
         $input->controllerExtends = $this->option('controller-extends') ?: null;
-        $input->withoutMigration = $this->option('without-migration');
+        $input->withoutMigration = $this->option('without-migration') || !Config::useCodeFirst();
         $input->force = $this->option('force');
         $input->modelDirectory = $this->option('model-directory');
         $input->primaryKey = $this->option('primary-key');